+++ frontend/package.json
@@ -15,6 +15,7 @@
     "@types/react": "^17.0.24",
     "@types/react-dom": "^17.0.9",
     "@types/react-redux": "^7.1.18",
+    "@types/react-router-dom": "^5.3.1",
     "axios": "^0.21.4",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
