+++ frontend/src/App.tsx
@@ -1,25 +1,19 @@
 import React from 'react';
+import { BrowserRouter, Switch, Route, Redirect } from 'react-router-dom';
 import './App.css';
-import { useAppSelector, useAppDispatch } from './redux/hooks'
-import { increment, incrementByAmount } from './redux/counter'
+
+import Home from './pages/Home';
+import Overview from './pages/Overview';
 
 function App() {
-
-  const count = useAppSelector((state) => state.counter.value)
-  const dispatch = useAppDispatch()
-
-  const onClick = () => {
-    dispatch(incrementByAmount(2))
-  }
-
   return (
-    <div className="App">
-      <header className="App-header">
-        <h2>{count}</h2>
-        <button onClick={onClick}>Increment by 2</button>
-        <button onClick={() => dispatch(increment())}>Increment by 1</button>
-      </header>
-    </div>
+    <BrowserRouter>
+      <Switch>
+        <Route exact path='/'><Home /></Route>
+        <Route exact path='/overview'><Overview /></Route>
+        <Redirect to='/' />
+      </Switch>
+    </BrowserRouter>
   );
 }
