+++ frontend/src/pages/Overview.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function Overview() {
+
+  return (
+    <header>
+      Placeholder Text (OVERVIEW)
+    </header>
+  );
+}
+
+export default Overview;
